+++ lib/mutant/mutator/node/noop.rb
@@ -34,6 +34,7 @@ module Mutant
         handle(Rubinius::AST::File)
         handle(Rubinius::AST::DynamicRegex)
         handle(Rubinius::AST::OpAssignOr19)
+        handle(Rubinius::AST::BlockPass19)
         handle(Rubinius::AST::OpAssign1)
         handle(Rubinius::AST::Or)
         handle(Rubinius::AST::ConstantAccess)