+++ spec/shared/mutator_behavior.rb
@@ -15,13 +15,25 @@ shared_examples_for 'a mutator' do
 
     it { should be_instance_of(to_enum.class) }
 
+    def assert_transitive(ast)
+      generated = ToSource.to_source(ast)
+      parsed    = generated.to_ast
+      again     = ToSource.to_source(parsed)
+      unless generated == again
+        fail "Untransitive:\n%s\n---\n%s" % [generated, again]
+      end
+    end
+
     unless instance_methods.include?(:expected_mutations)
       let(:expected_mutations)  do
         mutations.map do |mutation|
           case mutation
           when String
-            mutation.to_ast
+            ast = mutation.to_ast
+            assert_transitive(ast)
+            ast
           when Rubinius::AST::Node
+            assert_transitive(mutation)
             mutation
           else
             raise